poc/requirements.txt
# Multi-Provider Prompt Injection Testing System Requirements

# Core API provider dependencies
mistralai>=1.0.0
openai>=1.0.0
anthropic>=0.25.0

# Additional dependencies for enhanced functionality
requests>=2.25.0
aiohttp>=3.8.0

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
flake8>=5.0.0

# Data analysis and visualization (optional)
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and CLI enhancements (optional)
tqdm>=4.60.0
rich>=12.0.0

# Configuration management (optional)
python-dotenv>=0.19.0
pyyaml>=6.0

# Logging and monitoring (optional)
structlog>=22.0.0
